Writers: use longs in progress logs, fix #1137

+++ mapsforge-map-writer/src/main/java/org/mapsforge/map/writer/BaseTileBasedDataProcessor.java
@@ -2,7 +2,7 @@
  * Copyright 2010, 2011, 2012, 2013 mapsforge.org
  * Copyright 2015 lincomatic
  * Copyright 2017-2018 Gustl22
- * Copyright 2017 devemux86
+ * Copyright 2017-2019 devemux86
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -17,41 +17,6 @@
  */
 package org.mapsforge.map.writer;
 
-import org.locationtech.jts.geom.Polygon;
-import org.locationtech.jts.geom.TopologyException;
-import org.mapsforge.core.model.BoundingBox;
-import org.mapsforge.core.util.LatLongUtils;
-import org.mapsforge.core.util.MercatorProjection;
-import org.mapsforge.map.writer.model.MapWriterConfiguration;
-import org.mapsforge.map.writer.model.NodeResolver;
-import org.mapsforge.map.writer.model.OSMTag;
-import org.mapsforge.map.writer.model.TDNode;
-import org.mapsforge.map.writer.model.TDRelation;
-import org.mapsforge.map.writer.model.TDWay;
-import org.mapsforge.map.writer.model.TileBasedDataProcessor;
-import org.mapsforge.map.writer.model.TileCoordinate;
-import org.mapsforge.map.writer.model.TileData;
-import org.mapsforge.map.writer.model.TileGridLayout;
-import org.mapsforge.map.writer.model.TileInfo;
-import org.mapsforge.map.writer.model.WayResolver;
-import org.mapsforge.map.writer.model.ZoomIntervalConfiguration;
-import org.mapsforge.map.writer.util.GeoUtils;
-import org.mapsforge.map.writer.util.OSMUtils;
-import org.openstreetmap.osmosis.core.domain.v0_6.Node;
-import org.openstreetmap.osmosis.core.domain.v0_6.Relation;
-import org.openstreetmap.osmosis.core.domain.v0_6.Way;
-
-import java.text.NumberFormat;
-import java.util.ArrayList;
-import java.util.Deque;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.Set;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
 import gnu.trove.iterator.TLongIterator;
 import gnu.trove.list.array.TLongArrayList;
 import gnu.trove.map.hash.TLongObjectHashMap;
@@ -59,6 +24,23 @@ import gnu.trove.map.hash.TShortIntHashMap;
 import gnu.trove.procedure.TObjectProcedure;
 import gnu.trove.set.TLongSet;
 import gnu.trove.set.hash.TLongHashSet;
+import org.locationtech.jts.geom.Polygon;
+import org.locationtech.jts.geom.TopologyException;
+import org.mapsforge.core.model.BoundingBox;
+import org.mapsforge.core.util.LatLongUtils;
+import org.mapsforge.core.util.MercatorProjection;
+import org.mapsforge.map.writer.model.*;
+import org.mapsforge.map.writer.util.GeoUtils;
+import org.mapsforge.map.writer.util.OSMUtils;
+import org.openstreetmap.osmosis.core.domain.v0_6.Node;
+import org.openstreetmap.osmosis.core.domain.v0_6.Relation;
+import org.openstreetmap.osmosis.core.domain.v0_6.Way;
+
+import java.text.NumberFormat;
+import java.util.*;
+import java.util.Map.Entry;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 abstract class BaseTileBasedDataProcessor implements TileBasedDataProcessor, NodeResolver, WayResolver {
 
@@ -69,7 +51,7 @@ abstract class BaseTileBasedDataProcessor implements TileBasedDataProcessor, Nod
         private Map<Integer, List<Integer>> outerToInner;
         private final WayPolygonizer polygonizer = new WayPolygonizer();
 
-        private int nRelations = 0;
+        private long nRelations = 0;
 
         @Override
         public boolean execute(TDRelation relation) {
@@ -265,7 +247,7 @@ abstract class BaseTileBasedDataProcessor implements TileBasedDataProcessor, Nod
     }
 
     protected class WayHandler implements TObjectProcedure<TDWay> {
-        private int nWays = 0;
+        private long nWays = 0;
 
         @Override
         public boolean execute(TDWay way) {
@@ -323,9 +305,9 @@ abstract class BaseTileBasedDataProcessor implements TileBasedDataProcessor, Nod
     protected final Map<Long, Long> partRootRelations;
 
     // Accounting
-    private int amountOfNodesProcessed = 0;
-    private int amountOfRelationsProcessed = 0;
-    private int amountOfWaysProcessed = 0;
+    private long amountOfNodesProcessed = 0;
+    private long amountOfRelationsProcessed = 0;
+    private long amountOfWaysProcessed = 0;
     protected final NumberFormat nfCounts = NumberFormat.getInstance();
 
     protected final ZoomIntervalConfiguration zoomIntervalConfiguration;
@@ -426,17 +408,17 @@ abstract class BaseTileBasedDataProcessor implements TileBasedDataProcessor, Nod
     }
 
     @Override
-    public int getNodesNumber() {
+    public long getNodesNumber() {
         return amountOfNodesProcessed;
     }
 
     @Override
-    public int getRelationsNumber() {
+    public long getRelationsNumber() {
         return amountOfRelationsProcessed;
     }
 
     @Override
-    public int getWaysNumber() {
+    public long getWaysNumber() {
         return amountOfWaysProcessed;
     }
 
@@ -573,7 +555,7 @@ abstract class BaseTileBasedDataProcessor implements TileBasedDataProcessor, Nod
             return;
         }
 
-        /*int progressImplicitRelations = 0;
+        /*long progressImplicitRelations = 0;
         float limitImplicitRelations = this.tilesToPartElements.entrySet().size();*/
 
         // Iterate through tiles which contain parts
@@ -592,7 +574,7 @@ abstract class BaseTileBasedDataProcessor implements TileBasedDataProcessor, Nod
                     + "%% - Tile (" + tilePartElementEntry.getKey().getX()
                     + ", " + tilePartElementEntry.getKey().getY() + ")";
             progressImplicitRelations++;
-            int nRootElements = 0;*/
+            long nRootElements = 0;*/
 
             // Load parts only once in cache
             List<TDWay> pElems = new ArrayList<>();
